template args for space for values
